+++ sqlplus_session/__init__.py
@@ -29,7 +29,7 @@ from ._errors import (
     SqlplusTimeout,
 )
 
-__version__ = '0.4.0'
+__version__ = '0.5.0'
 
 __all__ = [
     'SqlplusSession',